Move NoSsr to unstyled package
