+++ docs/pages/api-docs/no-ssr.json
@@ -7,7 +7,7 @@
   "name": "NoSsr",
   "styles": { "classes": [], "globalClasses": {}, "name": null },
   "spread": false,
-  "filename": "/packages/material-ui/src/NoSsr/NoSsr.js",
+  "filename": "/packages/material-ui-unstyled/src/NoSsr/NoSsr.js",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/components/no-ssr/\">No Ssr</a></li></ul>",
   "styledComponent": true,
