+++ docs/pages/performance/slider-jss.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 import Slider from '@material-ui/core/Slider';
 
 const data = {
