+++ docs/pages/performance/table-emotion.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 // import { styled } from '@material-ui/core/styles';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 
 const createComponent = (defaultComponent) => {
   const Root = styled('div')`
