+++ docs/src/pages/customization/transitions/TransitionHover.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled, { ThemeProvider } from 'styled-components';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 import {
   createTheme,
   ThemeProvider as MuiThemeProvider,
