+++ docs/src/pages/styles/advanced/GlobalClassName.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { TextField, NoSsr } from '@material-ui/core';
+import { TextField } from '@material-ui/core';
+import { NoSsr } from '@material-ui/unstyled';
 
 const StyledTextField = styled(TextField)`
   label.Mui-focused {
