+++ docs/src/pages/system/advanced/CombiningStyleFunctionsDemo.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
 import { palette, PaletteProps, spacing, SpacingProps } from '@material-ui/system';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 
 const Div = styled.div<PaletteProps & SpacingProps>`
   ${palette}
