+++ docs/src/pages/system/advanced/StyleFunctionSxDemo.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled, { ThemeProvider } from 'styled-components';
 import { unstable_styleFunctionSx } from '@material-ui/system';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 import { createTheme } from '@material-ui/core/styles';
 
 const theme = createTheme();
