+++ docs/src/pages/system/advanced/StyleFunctionSxDemo.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled, { InterpolationFunction, ThemeProvider } from 'styled-components';
 import { unstable_styleFunctionSx, SxProps } from '@material-ui/system';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 import { createTheme } from '@material-ui/core/styles';
 
 interface DivProps {
