+++ packages/material-ui-unstyled/src/NoSsr/NoSsr.js
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { exactProp } from '@material-ui/utils';
-import useEnhancedEffect from '../utils/useEnhancedEffect';
+import { exactProp, unstable_useEnhancedEffect as useEnhancedEffect } from '@material-ui/utils';
 
 /**
  * NoSsr purposely removes components from the subject of Server Side Rendering (SSR).
