+++ packages/material-ui-unstyled/src/NoSsr/NoSsr.test.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { expect } from 'chai';
 import { createClientRender, createServerRender } from 'test/utils';
-import NoSsr from '@material-ui/core/NoSsr';
+import NoSsr from '@material-ui/unstyled/NoSsr';
 
 describe('<NoSsr />', () => {
   const render = createClientRender();
