+++ packages/material-ui-unstyled/src/NoSsr/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './NoSsr';
+export * from './NoSsr';
