+++ packages/material-ui-unstyled/src/NoSsr/index.js
@@ -1,0 +1,1 @@
+export { default } from './NoSsr';
