+++ packages/material-ui-unstyled/src/index.d.ts
@@ -4,20 +4,8 @@ export * from './BackdropUnstyled';
 export { default as BadgeUnstyled } from './BadgeUnstyled';
 export * from './BadgeUnstyled';
 
-export { default as ModalUnstyled } from './ModalUnstyled';
-export * from './ModalUnstyled';
-
-export { default as SliderUnstyled } from './SliderUnstyled';
-export * from './SliderUnstyled';
-
-export { default as SwitchUnstyled } from './SwitchUnstyled';
-export * from './SwitchUnstyled';
-
-export { default as Portal } from './Portal';
-export * from './Portal';
-
-export { default as Unstable_TrapFocus } from './Unstable_TrapFocus';
-export * from './Unstable_TrapFocus';
+export { default as unstable_composeClasses } from './composeClasses';
+export * from './composeClasses';
 
 export { default as generateUtilityClass } from './generateUtilityClass';
 export * from './generateUtilityClass';
@@ -25,7 +13,21 @@ export * from './generateUtilityClass';
 export { default as generateUtilityClasses } from './generateUtilityClasses';
 export * from './generateUtilityClasses';
 
-export { default as unstable_composeClasses } from './composeClasses';
-export * from './composeClasses';
+export { default as ModalUnstyled } from './ModalUnstyled';
+export * from './ModalUnstyled';
+
+export { default as NoSsr } from './NoSsr';
+
+export { default as Portal } from './Portal';
+export * from './Portal';
+
+export { default as SliderUnstyled } from './SliderUnstyled';
+export * from './SliderUnstyled';
+
+export { default as SwitchUnstyled } from './SwitchUnstyled';
+export * from './SwitchUnstyled';
+
+export { default as Unstable_TrapFocus } from './Unstable_TrapFocus';
+export * from './Unstable_TrapFocus';
 
 export * from './utils';
