+++ packages/material-ui-unstyled/src/index.js
@@ -4,23 +4,25 @@ export * from './BackdropUnstyled';
 export { default as BadgeUnstyled } from './BadgeUnstyled';
 export * from './BadgeUnstyled';
 
+export { default as unstable_composeClasses } from './composeClasses';
+
+export { default as generateUtilityClass } from './generateUtilityClass';
+
+export { default as generateUtilityClasses } from './generateUtilityClasses';
+
+export { default as ModalUnstyled } from './ModalUnstyled';
+export * from './ModalUnstyled';
+
+export { default as NoSsr } from './NoSsr';
+
+export { default as Portal } from './Portal';
+
 export { default as SliderUnstyled } from './SliderUnstyled';
 export * from './SliderUnstyled';
 
 export { default as SwitchUnstyled } from './SwitchUnstyled';
 export * from './SwitchUnstyled';
 
-export { default as ModalUnstyled } from './ModalUnstyled';
-export * from './ModalUnstyled';
-
-export { default as Portal } from './Portal';
-
 export { default as Unstable_TrapFocus } from './Unstable_TrapFocus';
-
-export { default as generateUtilityClass } from './generateUtilityClass';
-
-export { default as generateUtilityClasses } from './generateUtilityClasses';
-
-export { default as unstable_composeClasses } from './composeClasses';
 
 export * from './utils';
