+++ packages/material-ui/src/NoSsr/index.js
@@ -1,1 +1,1 @@
-export { default } from './NoSsr';
+export { default } from '@material-ui/unstyled/NoSsr';
